修改逻辑bugfix

+++ src/main/java/com/opzpy123/controller/BlogController.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
+import java.security.Principal;
 import java.util.List;
 
 @Slf4j
@@ -28,23 +29,31 @@ public class BlogController {
     @PostMapping
     @ResponseBody
     ApiResponse<String> addBlog(@RequestBody Blog blog) {
-        log.info(blog.getUserName()+"创建了blog:"+blog.getTitle());
+        log.info(blog.getUserName() + "创建了blog:" + blog.getTitle());
         return blogService.addBlog(blog);
     }
 
     @PutMapping
     @ResponseBody
     ApiResponse<String> updateBlog(@RequestBody Blog blog) {
+        blog.setUserName(null);
         blogService.updateById(blog);
-        log.info(blog.getUserName()+"修改了blog:"+blog.getTitle());
+        log.info(blog.getUserName() + "修改了blog:" + blog.getTitle());
         return ApiResponse.ofSuccess();
     }
 
     @DeleteMapping
     @ResponseBody
-    ApiResponse<String> deleteBlog(@RequestBody Blog blog) {
-        blogService.removeById(blog);
-        log.info(blog.getUserName()+"删除了blog:"+blog.getTitle());
+    ApiResponse<String> deleteBlog(@RequestBody Blog blog, Principal principal) {
+        if (principal.getName().equals("admin")) {
+            blogService.removeById(blog);
+        } else {
+            Blog blogById = blogService.getById(blog.getId());
+            if (blogById != null && blogById.getUserName().equals(principal.getName())) {
+                blogService.removeById(blog);
+            }
+        }
+        log.info(blog.getUserName() + "删除了blog:" + blog.getTitle());
         return ApiResponse.ofSuccess();
     }
 }
